requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
requests==2.28.1
python-dotenv==0.21.0
flask_sqlalchemy == 3.0.2
# Flask-Login==0.6.2
psycopg2-binary == 2.9.5
# psycopg2 == 2.9.5
